Reworked module set up to attempt to fix non-running spec..
